refactor(chat): isolate transcript content reads
